decimal internal transfer

+++ src/Service.InternalTransfer.Grpc/Models/TransferByPhoneRequest.cs
@@ -10,7 +10,7 @@ namespace Service.InternalTransfer.Grpc.Models
         [DataMember(Order = 3)] public string ClientId { get; set; }
         [DataMember(Order = 4)] public string WalletId { get; set; }
         [DataMember(Order = 5)] public string AssetSymbol { get; set; }
-        [DataMember(Order = 6)] public double Amount { get; set; }
+        [DataMember(Order = 6)] public decimal Amount { get; set; }
         [DataMember(Order = 7)] public string ToPhoneNumber { get; set; }
         [DataMember(Order = 8)] public string ClientLang { get; set; }
         [DataMember(Order = 9)] public string ClientIp { get; set; }
